refactor(divergence): split target evidence modules
