+++ crates/nose-cli/src/query_views.rs
@@ -1,3 +1,5 @@
+mod base_json;
+
 use super::query_model::*;
 use crate::baseline;
 use crate::baseline_comparison::BaselineComparison;
@@ -99,7 +101,7 @@ pub(super) fn render_query_base(
     let fire_eligible = flagged.iter().filter(|d| d.fire_eligible).count();
     let strict = flagged.iter().filter(|d| d.gate_fail_default()).count();
     if json {
-        render_query_base_json(flagged, changed_files, base_ref, path, top, semantic_packs);
+        base_json::render(flagged, changed_files, base_ref, path, top, semantic_packs);
         return;
     }
     print_query_prelude();
@@ -174,49 +176,6 @@ pub(super) fn render_query_base(
     }
     println!("\nnext:");
     println!("  nose query {path} base={base_ref} --fail-on any   # fail CI on strict divergences");
-}
-
-fn render_query_base_json(
-    flagged: &[divergence::Divergence],
-    changed_files: usize,
-    base_ref: &str,
-    path: &str,
-    top: Option<usize>,
-    semantic_packs: &[serde_json::Value],
-) {
-    let limit = query_row_limit(top);
-    let items: Vec<_> = divergence::divergence_items_json(flagged)
-        .into_iter()
-        .take(limit)
-        .collect();
-    let limit_value = match top {
-        Some(0) => serde_json::Value::Null,
-        Some(n) => serde_json::json!(n),
-        None => serde_json::json!(30),
-    };
-    println!(
-        "{}",
-        with_semantic_packs(
-            serde_json::json!({
-                "schema_version": schema_versions::QUERY_BASE_JSON_SCHEMA_VERSION,
-                "tool": "nose",
-                "view": "base",
-                "path": path,
-                "base": base_ref,
-                "summary": {
-                    "changed_files": changed_files,
-                    "divergences": flagged.len(),
-                    "shown_divergences": items.len(),
-                    "limit": limit_value,
-                    "fire_eligible": flagged.iter().filter(|d| d.fire_eligible).count(),
-                    "strict": flagged.iter().filter(|d| d.gate_fail_default()).count(),
-                },
-                "items": items,
-                "next": [format!("nose query {path} base={base_ref} --fail-on any")],
-            }),
-            semantic_packs
-        )
-    );
 }
 
 fn print_semantic_change(site: &divergence::Site) {
